Make strong-remoting-android version configurable
Modify build.gradle and gradle.properties so that it's possible to edit
version of strong-remoting-android when building a release. This way
the CI build can use the latest SNAPSHOT dependency and it's still possible
to use a fixed version for a release build.

See this StackOverflow comment for more details:
http://stackoverflow.com/questions/4407055/how-do-i-stop-maven-version-ranges-from-using-snapshots#comment18858069_4407055